requirements.txt
nltk
gensim
pdfminer.six
scikit-learn
symspellpy
beautifulsoup4